First Change -- IDE Color
With my poor eyes, this is *the* essential first change necessary for SDL -- changing the text from the faded gray to the bright white so that it shows up and is readable to me in the IDE, without any eye strain.  Other stuff comes later; this change is essential.  :D